Go 1.4: Move crypto dependency to golang.org
Now that the /x/crypto package has been moved to git, we can reference it
with a git submodule, instead of the old way we were using, which was to
commit literal copies of all the source files from the Mercurial-based
dependencies.
The only remaining Mercurial based dependency is winsvc.

This change was mandated by us upgrading winsvc, which in turn forced us
to upgrade /x/crypto.